HW4 FIX 1 Reason: code review
