+++ react/src/components/menu/MenuItem.jsx
@@ -1,4 +1,4 @@
-import { Counter } from "../counter/Counter.jsx";
+import { DishCounter } from "../counters/DishCounter.jsx";
 
 const TOPIC_ORDER = "Order";
 
@@ -12,7 +12,7 @@ export const MenuItem = ({ itemName, menuItems }) => {
             <h3>{name}</h3>
             <p>{ingredients.join(', ')}</p>
             <p>£{price}</p>
-            <Counter topic={TOPIC_ORDER}/>
+            <DishCounter topic={TOPIC_ORDER}/>
           </li>
         ))}
       </ul>
